# chore(client): Performance improvements on initial sync downstream path