--- a/.changeset/beige-lemons-move.md
+++ b/.changeset/beige-lemons-move.md
@@ -1,0 +1,5 @@
+---
+"electric-sql": patch
+---
+
+Minor performance improvements in downstream initial sync path