@alco 2024-06-14
> There are ways we could optimise this path but I was reluctant to spend time on that

I expect that we'll see a lot of such loose threads of optimisation. No good jumping on each of them immediately, but we should keep track of those someplace to be able to evaluate later which potential optimisation would have the biggest impact.